+++ bico/misc/randomness.cpp
@@ -2,4 +2,4 @@
 
 using namespace CluE;
 
-std::mt19937 Randomness::mt19937Generator(time(0));
+std::mt19937 Randomness::mt19937Generator(static_cast<unsigned int>(time(0)));
